+++ examples/basic_window.rs
@@ -1,130 +1,160 @@
-extern crate xcb;
+use std::str;
+use xcb::{x, Xid};
 
-use std::iter::Iterator;
-
-fn main() {
+fn main() -> xcb::Result<()> {
     let (conn, screen_num) = xcb::Connection::connect(None).unwrap();
     let setup = conn.get_setup();
     let screen = setup.roots().nth(screen_num as usize).unwrap();
 
-    let window = conn.generate_id();
+    let window: x::Window = conn.generate_id();
 
-    let values = [
-        (xcb::CW_BACK_PIXEL, screen.white_pixel()),
-        (
-            xcb::CW_EVENT_MASK,
-            xcb::EVENT_MASK_EXPOSURE | xcb::EVENT_MASK_KEY_PRESS,
-        ),
-    ];
+    println!("window: {}", window.resource_id());
+    println!("root: {}", screen.root().resource_id());
+    println!("root_visual: {}", screen.root_visual());
 
-    xcb::create_window(
-        &conn,
-        xcb::COPY_FROM_PARENT as u8,
-        window,
-        screen.root(),
-        0,
-        0,
-        150,
-        150,
-        10,
-        xcb::WINDOW_CLASS_INPUT_OUTPUT as u16,
-        screen.root_visual(),
-        &values,
-    );
+    conn.send_request(&x::CreateWindow {
+        depth: x::COPY_FROM_PARENT as u8,
+        wid: window,
+        parent: screen.root(),
+        x: 0,
+        y: 0,
+        width: 150,
+        height: 150,
+        border_width: 10,
+        class: x::WindowClass::InputOutput,
+        visual: screen.root_visual(),
+        value_list: &[
+            x::Cw::BackPixel(screen.white_pixel()),
+            x::Cw::EventMask((x::EventMask::EXPOSURE | x::EventMask::KEY_PRESS).bits()),
+        ],
+    });
 
-    xcb::map_window(&conn, window);
+    conn.send_request(&x::MapWindow { window });
 
     let title = "Basic Window";
-    // setting title
-    xcb::change_property(
-        &conn,
-        xcb::PROP_MODE_REPLACE as u8,
-        window,
-        xcb::ATOM_WM_NAME,
-        xcb::ATOM_STRING,
-        8,
-        title.as_bytes(),
-    );
 
-    conn.flush();
+    conn.send_request(&x::ChangeProperty {
+        mode: x::PropMode::Replace,
+        window,
+        property: x::ATOM_WM_NAME,
+        r#type: x::ATOM_STRING,
+        data: title.as_bytes(),
+    });
+
+    conn.flush()?;
 
     // retrieving title
-    let cookie = xcb::get_property(
-        &conn,
-        false,
+    let cookie = conn.send_request(&x::GetProperty {
+        delete: false,
         window,
-        xcb::ATOM_WM_NAME,
-        xcb::ATOM_STRING,
-        0,
-        1024,
-    );
-    if let Ok(reply) = cookie.get_reply() {
-        assert_eq!(std::str::from_utf8(reply.value()).unwrap(), title);
-    } else {
-        panic!("could not retrieve window title!");
-    }
+        property: x::ATOM_WM_NAME,
+        r#type: x::ATOM_STRING,
+        long_offset: 0,
+        long_length: 1024,
+    });
+    let reply = conn.wait_for_reply(cookie)?;
+    assert_eq!(str::from_utf8(reply.value()).unwrap(), title);
 
     // retrieving a few atoms
-    let (wm_state, wm_state_maxv, wm_state_maxh) = {
-        let cook = xcb::intern_atom(&conn, true, "_NET_WM_STATE");
-        let cook_maxv = xcb::intern_atom(&conn, true, "_NET_WM_STATE_MAXIMIZED_VERT");
-        let cook_maxh = xcb::intern_atom(&conn, true, "_NET_WM_STATE_MAXIMIZED_HORZ");
-
+    let (wm_protocols, wm_del_window, wm_state, wm_state_maxv, wm_state_maxh) = {
+        let cookies = (
+            conn.send_request(&x::InternAtom {
+                only_if_exists: true,
+                name: "WM_PROTOCOLS",
+            }),
+            conn.send_request(&x::InternAtom {
+                only_if_exists: true,
+                name: "WM_DELETE_WINDOW",
+            }),
+            conn.send_request(&x::InternAtom {
+                only_if_exists: true,
+                name: "_NET_WM_STATE",
+            }),
+            conn.send_request(&x::InternAtom {
+                only_if_exists: true,
+                name: "_NET_WM_STATE_MAXIMIZED_VERT",
+            }),
+            conn.send_request(&x::InternAtom {
+                only_if_exists: true,
+                name: "_NET_WM_STATE_MAXIMIZED_HORZ",
+            }),
+        );
         (
-            cook.get_reply().unwrap().atom(),
-            cook_maxv.get_reply().unwrap().atom(),
-            cook_maxh.get_reply().unwrap().atom(),
+            conn.wait_for_reply(cookies.0)?.atom(),
+            conn.wait_for_reply(cookies.1)?.atom(),
+            conn.wait_for_reply(cookies.2)?.atom(),
+            conn.wait_for_reply(cookies.3)?.atom(),
+            conn.wait_for_reply(cookies.4)?.atom(),
         )
     };
+
+    // activate the sending of close event through `x::Event::ClientMessage`
+    // either the request must be checked as follow, or conn.flush() must be called before entering the loop
+    conn.check_request(conn.send_request_checked(&x::ChangeProperty {
+        mode: x::PropMode::Replace,
+        window,
+        property: wm_protocols,
+        r#type: x::ATOM_ATOM,
+        data: &[wm_del_window],
+    }))?;
 
     let mut maximized = false;
 
     loop {
-        let event = conn.wait_for_event();
-        match event {
-            None => {
-                break;
+        let event = match conn.wait_for_event() {
+            Err(xcb::Error::Connection(err)) => {
+                panic!("unexpected I/O error: {}", err);
             }
-            Some(event) => {
-                let r = event.response_type();
-                if r == xcb::KEY_PRESS as u8 {
-                    let key_press: &xcb::KeyPressEvent = unsafe { xcb::cast_event(&event) };
+            Err(xcb::Error::Protocol(err)) => {
+                panic!("unexpected protocol error: {:#?}", err);
+            }
+            Ok(event) => event,
+        };
 
-                    println!("Key '{}' pressed", key_press.detail());
+        println!("Received event {:#?}", event);
+        match event {
+            xcb::Event::X(x::Event::KeyPress(ev)) => {
+                println!("Key '{}' pressed", ev.detail());
 
-                    if key_press.detail() == 0x3a {
-                        // M (on qwerty)
+                if ev.detail() == 0x3a {
+                    // M (on qwerty)
 
-                        // toggle maximized
+                    // toggle maximized
 
-                        // ClientMessageData is a memory safe untagged union
-                        let data = xcb::ClientMessageData::from_data32([
-                            if maximized { 0 } else { 1 },
-                            wm_state_maxv,
-                            wm_state_maxh,
-                            0,
-                            0,
-                        ]);
+                    let data = x::ClientMessageData::Data32([
+                        if maximized { 0 } else { 1 },
+                        wm_state_maxv.resource_id(),
+                        wm_state_maxh.resource_id(),
+                        0,
+                        0,
+                    ]);
 
-                        let ev = xcb::ClientMessageEvent::new(32, window, wm_state, data);
+                    let event = x::ClientMessageEvent::new(window, wm_state, data);
 
-                        xcb::send_event(
-                            &conn,
-                            false,
-                            screen.root(),
-                            xcb::EVENT_MASK_STRUCTURE_NOTIFY,
-                            &ev,
-                        );
+                    conn.send_request(&x::SendEvent {
+                        propagate: false,
+                        destination: x::SendEventDest::Window(screen.root()),
+                        event_mask: x::EventMask::STRUCTURE_NOTIFY,
+                        event: &event,
+                    });
 
-                        conn.flush();
+                    conn.flush()?;
 
-                        maximized = !maximized;
-                    } else if key_press.detail() == 0x18 {
-                        // Q (on qwerty)
-                        break;
+                    maximized = !maximized;
+                } else if ev.detail() == 0x18 {
+                    // Q (on qwerty)
+                    break Ok(());
+                }
+            }
+            xcb::Event::X(x::Event::ClientMessage(ev)) => {
+                if let x::ClientMessageData::Data32([atom, ..]) = ev.data() {
+                    if atom == wm_del_window.resource_id() {
+                        // window "x" button clicked by user, we gracefully exit
+                        break Ok(());
                     }
                 }
             }
+            _ => {}
         }
     }
 }