+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased-WIP]
+- new safer API
+  - implement the extensions in RUST directly
+  - safe enums for Errors and Events
+  - requests are structures that can be passed to `Connection::send_request`
+  - checked/unchecked and void/non-void requests are type safe (each request has two cookie type)
+- support for xcb-1.14
+- support for xinput extension
+
 ## [0.10.1] - 2021-09-23 - rtbo
 - fix some code generation affecting the `present` extension
 - fix compilation warnings about uninhabited type values